Do not require ws feature for low-level APIs
