+++ kube/src/api/subresource.rs
@@ -3,7 +3,7 @@ use futures::Stream;
 use serde::de::DeserializeOwned;
 
 use crate::{
-    api::{Api, DeleteParams, Patch, PatchParams, PostParams, Resource},
+    api::{Api, AttachParams, DeleteParams, Patch, PatchParams, PostParams, Resource},
     client::Status,
     Error, Result,
 };
@@ -288,199 +288,6 @@ where
 // ----------------------------------------------------------------------------
 // Attach subresource
 // ----------------------------------------------------------------------------
-/// Parameters for attaching to a container in a Pod.
-///
-/// - One of `stdin`, `stdout`, or `stderr` must be `true`.
-/// - `stderr` and `tty` cannot both be `true` because multiplexing is not supported with TTY.
-#[cfg(feature = "ws")]
-#[cfg_attr(docsrs, doc(cfg(feature = "ws")))]
-pub struct AttachParams {
-    /// The name of the container to attach.
-    /// Defaults to the only container if there is only one container in the pod.
-    pub container: Option<String>,
-    /// Attach to the container's standard input. Defaults to `false`.
-    ///
-    /// Call [`AttachedProcess::stdin`] to obtain a writer.
-    pub stdin: bool,
-    /// Attach to the container's standard output. Defaults to `true`.
-    ///
-    /// Call [`AttachedProcess::stdout`] to obtain a reader.
-    pub stdout: bool,
-    /// Attach to the container's standard error. Defaults to `true`.
-    ///
-    /// Call [`AttachedProcess::stderr`] to obtain a reader.
-    pub stderr: bool,
-    /// Allocate TTY. Defaults to `false`.
-    ///
-    /// NOTE: Terminal resizing is not implemented yet.
-    pub tty: bool,
-
-    /// The maximum amount of bytes that can be written to the internal `stdin`
-    /// pipe before the write returns `Poll::Pending`.
-    /// Defaults to 1024.
-    ///
-    /// This is not sent to the server.
-    pub max_stdin_buf_size: Option<usize>,
-    /// The maximum amount of bytes that can be written to the internal `stdout`
-    /// pipe before the write returns `Poll::Pending`.
-    /// Defaults to 1024.
-    ///
-    /// This is not sent to the server.
-    pub max_stdout_buf_size: Option<usize>,
-    /// The maximum amount of bytes that can be written to the internal `stderr`
-    /// pipe before the write returns `Poll::Pending`.
-    /// Defaults to 1024.
-    ///
-    /// This is not sent to the server.
-    pub max_stderr_buf_size: Option<usize>,
-}
-
-#[cfg(feature = "ws")]
-impl Default for AttachParams {
-    // Default matching the server's defaults.
-    fn default() -> Self {
-        Self {
-            container: None,
-            stdin: false,
-            stdout: true,
-            stderr: true,
-            tty: false,
-            max_stdin_buf_size: None,
-            max_stdout_buf_size: None,
-            max_stderr_buf_size: None,
-        }
-    }
-}
-
-#[cfg(feature = "ws")]
-impl AttachParams {
-    /// Default parameters for an tty exec with stdin and stdout
-    pub fn interactive_tty() -> Self {
-        Self {
-            stdin: true,
-            stdout: true,
-            stderr: false,
-            tty: true,
-            ..Default::default()
-        }
-    }
-
-    /// Specify the container to execute in.
-    pub fn container<T: Into<String>>(mut self, container: T) -> Self {
-        self.container = Some(container.into());
-        self
-    }
-
-    /// Set `stdin` field.
-    pub fn stdin(mut self, enable: bool) -> Self {
-        self.stdin = enable;
-        self
-    }
-
-    /// Set `stdout` field.
-    pub fn stdout(mut self, enable: bool) -> Self {
-        self.stdout = enable;
-        self
-    }
-
-    /// Set `stderr` field.
-    pub fn stderr(mut self, enable: bool) -> Self {
-        self.stderr = enable;
-        self
-    }
-
-    /// Set `tty` field.
-    pub fn tty(mut self, enable: bool) -> Self {
-        self.tty = enable;
-        self
-    }
-
-    /// Set `max_stdin_buf_size` field.
-    pub fn max_stdin_buf_size(mut self, size: usize) -> Self {
-        self.max_stdin_buf_size = Some(size);
-        self
-    }
-
-    /// Set `max_stdout_buf_size` field.
-    pub fn max_stdout_buf_size(mut self, size: usize) -> Self {
-        self.max_stdout_buf_size = Some(size);
-        self
-    }
-
-    /// Set `max_stderr_buf_size` field.
-    pub fn max_stderr_buf_size(mut self, size: usize) -> Self {
-        self.max_stderr_buf_size = Some(size);
-        self
-    }
-
-    fn validate(&self) -> Result<()> {
-        if !self.stdin && !self.stdout && !self.stderr {
-            return Err(Error::RequestValidation(
-                "AttachParams: one of stdin, stdout, or stderr must be true".into(),
-            ));
-        }
-
-        if self.stderr && self.tty {
-            // Multiplexing is not supported with TTY
-            return Err(Error::RequestValidation(
-                "AttachParams: tty and stderr cannot both be true".into(),
-            ));
-        }
-
-        Ok(())
-    }
-
-    fn append_to_url_serializer(&self, qp: &mut url::form_urlencoded::Serializer<String>) {
-        if self.stdin {
-            qp.append_pair("stdin", "true");
-        }
-        if self.stdout {
-            qp.append_pair("stdout", "true");
-        }
-        if self.stderr {
-            qp.append_pair("stderr", "true");
-        }
-        if self.tty {
-            qp.append_pair("tty", "true");
-        }
-        if let Some(container) = &self.container {
-            qp.append_pair("container", &container);
-        }
-    }
-}
-
-#[cfg(feature = "ws")]
-#[cfg_attr(docsrs, doc(cfg(feature = "ws")))]
-impl Resource {
-    /// Attach to a pod
-    pub fn attach(&self, name: &str, ap: &AttachParams) -> Result<http::Request<Vec<u8>>> {
-        ap.validate()?;
-
-        let base_url = self.make_url() + "/" + name + "/" + "attach?";
-        let mut qp = url::form_urlencoded::Serializer::new(base_url);
-        ap.append_to_url_serializer(&mut qp);
-
-        let req = http::Request::get(qp.finish());
-        req.body(vec![]).map_err(Error::HttpError)
-    }
-}
-
-#[cfg(feature = "ws")]
-#[test]
-fn attach_path() {
-    use crate::api::Resource;
-    use k8s_openapi::api::core::v1 as corev1;
-    let r = Resource::namespaced::<corev1::Pod>("ns");
-    let ap = AttachParams {
-        container: Some("blah".into()),
-        ..AttachParams::default()
-    };
-    let req = r.attach("foo", &ap).unwrap();
-    assert_eq!(
-        req.uri(),
-        "/api/v1/namespaces/ns/pods/foo/attach?&stdout=true&stderr=true&container=blah"
-    );
-}
 
 /// Marker trait for objects that has attach
 #[cfg(feature = "ws")]
@@ -508,46 +315,6 @@ where
 // ----------------------------------------------------------------------------
 // Exec subresource
 // ----------------------------------------------------------------------------
-#[cfg(feature = "ws")]
-#[cfg_attr(docsrs, doc(cfg(feature = "ws")))]
-impl Resource {
-    /// Execute command in a pod
-    pub fn exec<I, T>(&self, name: &str, command: I, ap: &AttachParams) -> Result<http::Request<Vec<u8>>>
-    where
-        I: IntoIterator<Item = T>,
-        T: Into<String>,
-    {
-        ap.validate()?;
-
-        let base_url = self.make_url() + "/" + name + "/" + "exec?";
-        let mut qp = url::form_urlencoded::Serializer::new(base_url);
-        ap.append_to_url_serializer(&mut qp);
-
-        for c in command.into_iter() {
-            qp.append_pair("command", &c.into());
-        }
-
-        let req = http::Request::get(qp.finish());
-        req.body(vec![]).map_err(Error::HttpError)
-    }
-}
-
-#[cfg(feature = "ws")]
-#[test]
-fn exec_path() {
-    use crate::api::Resource;
-    use k8s_openapi::api::core::v1 as corev1;
-    let r = Resource::namespaced::<corev1::Pod>("ns");
-    let ap = AttachParams {
-        container: Some("blah".into()),
-        ..AttachParams::default()
-    };
-    let req = r.exec("foo", vec!["echo", "foo", "bar"], &ap).unwrap();
-    assert_eq!(
-        req.uri(),
-        "/api/v1/namespaces/ns/pods/foo/exec?&stdout=true&stderr=true&container=blah&command=echo&command=foo&command=bar"
-    );
-}
 
 /// Marker trait for objects that has exec
 #[cfg(feature = "ws")]
